backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
itsdangerous
requests